backend/requirements.txt
# Install inside a venv only: python -m venv venv && .\venv\Scripts\activate
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
mysqlclient>=2.2
python-dotenv>=1.0
pandas>=2.0
numpy>=1.26
waitress>=3.0
whitenoise>=6.6
